requirements.txt
requests==2.26.0
telegram_send==0.25
mysql_connector_repackaged==0.3.1
python-dotenv==0.19.1
